requirements.txt
ollama
rich
arxiv
requests
xmltodict
pydantic
pyyaml